Added basic devise support
Added devise gem and implementation for user authentication including
adding user DSL to automatically make users editable.

The user model is created in the AppKit namespace. This required
refactoring of the way models are handled to allow for namespacing in
models

Support is needed for authorization and privlege management still.