Refactored table partial and changed resource index page to use it
The AppKit::Views::Table partial is now more broad and the index page
for resource views was changed to also use this partial. This most of
the tables to use the same code.

The association tables in the #show actions for resources still has its
own custom table code. For the moment this is due to future need to add
additional logic for nested resource creation. This may also be
refactored to use the table partial and object later though.