Classified configuration and refactored AppKit::Application
The application configuration is no longer a simple hash. It is now a
full object.

The Application class has been moved from being used as a constant in
the AppKit::Module to a class variable in AppKit. Because of the dynamic
nature of the class it seemed for fitting to make it a class variable
instead of a constant.